Seperate logging feature out into Rotoscope::CallLogger
The logging code seems out of place when using the Rotoscope
class with a custom trace proc.
