+++ ext/rotoscope/rotoscope.c
@@ -293,7 +293,7 @@ void Init_rotoscope(void) {
 
   cRotoscope = rb_define_class("Rotoscope", rb_cObject);
   rb_define_alloc_func(cRotoscope, rs_alloc);
-  rb_define_method(cRotoscope, "initialize_ext", rotoscope_initialize, 0);
+  rb_define_method(cRotoscope, "initialize", rotoscope_initialize, 0);
   rb_define_method(cRotoscope, "start_trace", rotoscope_start_trace, 0);
   rb_define_method(cRotoscope, "stop_trace", rotoscope_stop_trace, 0);
   rb_define_method(cRotoscope, "tracing?", rotoscope_tracing_p, 0);
